Keep spyre topk index in its input dtype through the decomposition
The device topk reduction materializes the index in the input dtype (a
float that lies it is an index), not int64. Teach Dynamo this by having
spyre::topkindex's fake report the input dtype, so it matches the real
lowering (lower_topkindex uses dst_dtype = x.get_dtype()) and the meta
does not conflict with the decomposition. The spyre_topk decomposition
returns the topkindex output directly, and the k ceiling is lifted from
4 to 128 (the work-division-across-cores limit).

+++ torch_spyre/_inductor/customops.py
@@ -151,7 +151,9 @@ def _(x: torch.Tensor, k: int, dim: int) -> torch.Tensor:
     norm_dim = dim % len(x.size())
     out_size = list(x.size())
     out_size[norm_dim] = k
-    return x.new_empty(out_size, dtype=torch.int64)
+    # Index materializes in the input dtype, not int64: a float that lies it
+    # is an index. Matches lower_topkindex (dst_dtype = x.get_dtype()).
+    return x.new_empty(out_size, dtype=x.dtype)
 
 
 @torch.library.custom_op("spyre::gelu", mutates_args=(), device_types="spyre")
